+++ src/Plainion.GraphViz.Modules.CodeInspection/Packaging/Spec/Cluster.cs
@@ -3,13 +3,15 @@ using System.Windows.Markup;
 
 namespace Plainion.GraphViz.Modules.CodeInspection.Packaging.Spec
 {
-    [ContentProperty( "Patterns" )]
+    [ContentProperty("Patterns")]
     public class Cluster : PackageBase
     {
         public string Id { get; set; }
 
-        internal bool Matches( string str )
+        internal bool Matches(string str)
         {
+            Contract.RequiresNotNull(str, nameof(str));
+
             return Includes.Any(i => i.Matches(str)) && !Excludes.Any(e => e.Matches(str));
         }
     }